refactor: cleaning up the TagBadgeCommon module
