--- packages/bezier-react/src/components/TagBadge/TagBadgeCommon/constants/TagBadgeSize.ts
@@ -1,8 +1,0 @@
-enum TagBadgeSize {
-  XS = 'xs',
-  S = 's',
-  M = 'm',
-  L = 'l',
-}
-
-export default TagBadgeSize
