--- packages/bezier-react/src/components/TagBadge/TagBadgeCommon/constants/TagBadgeStyle.ts
@@ -1,3 +1,0 @@
-import { IconSize } from '~/src/components/Icon'
-
-export const TAG_BADGE_ICON_SIZE = IconSize.XS
