+++ packages/bezier-react/src/components/TagBadge/TagBadgeCommon/index.ts
@@ -1,16 +1,1 @@
-/* Constants */
-export { default as TagBadgeSize } from './constants/TagBadgeSize'
-export {
-  default as TagBadgeVariant,
-  TagBadgeBgColorPreset,
-  BadgeColorPreset,
-} from './constants/TagBadgeVariant'
-export {
-  TAG_BADGE_ICON_SIZE,
-} from './constants/TagBadgeStyle'
-
-/* Utils */
-export {
-  getProperTagBadgeBgColor,
-  getProperTagBadgeTypo,
-} from './utils'
+export * from './TagBadgeCommon'
